added CRUD for accommodations and amenities.

+++ src/main/java/italy/accommodations/entity/Accommodation.java
@@ -32,7 +32,7 @@ public class Accommodation {
 	private String accommodationAddress;
 	
 	private String accommodationPhoneNumber;
-	private Double accommodationRating;
+	private int accommodationRating;
 	
 	
 	@EqualsAndHashCode.Exclude
